requirements.txt
requests
pyyaml
dnspython
termcolor
